refactor: seniorProfile 구조 조정
