+++ src/pages/seniorProfile/components/Init.tsx
@@ -1,7 +1,7 @@
 import { FullBtn } from '@components/commons/FullButton';
-import { funnelComponentPropType } from '@pages/seniorProfile/types';
 
-const Init = ({ setStep }: funnelComponentPropType) => {
+// eslint-disable-next-line no-undef
+const Init = ({ setStep }: { setStep: React.Dispatch<React.SetStateAction<number>> }) => {
   return <FullBtn text="다음으로" onClick={() => setStep && setStep((prev) => prev + 1)} isActive={true} />;
 };
